--- app/static/css/index.css
@@ -1,176 +1,0 @@
-/*main page*/
-.intro-text{
-    text-align: center;
-    font-family: Roboto, sans-serif;
-}
-div.container-fluid.text-white.pt-5 > div.row.tags > div > button:hover,
-div.container-fluid.text-white.pt-5 > div.row.tags > div > button:focus,
-div.container-fluid.text-white.pt-5 > div.row.tags > div > button:active{
-    background-color: #262b35;
-    color: #ffffff;
-}
-/*large category button*/
-div.container-fluid.text-white.pt-5 > div.row.tags > div> button{
-border-radius: 1.3rem;
-    border: none;
-    font-size: 1.1rem;
-    background-color: #262b35;
-    color: #ffffff;
-}
-div.container-fluid.text-white.pt-5 > div.row.justify-content-md-center.tags > div > button:hover,
-div.container-fluid.text-white.pt-5 > div.row.justify-content-md-center.tags > div > button:focus,
-div.container-fluid.text-white.pt-5 > div.row.justify-content-md-center.tags > div > button:active{
-    background-color: #ffffff;
-    border-radius: 1.3rem;
-    border: none;
-    font-size: 1.1rem;
-    color: #424242;
-}
-/*small category button*/
-div.container-fluid.text-white.pt-5 > div.row.justify-content-md-center.tags{
-    text-align: center;
-}
-#flexCheckDefault{
-    background-color: #e3b110;
-    border: none;
-}
-#flexCheckDefault:checked{
-    background-color: #1f284a;
-    border: none;
-}
-.form-check-label{
-    color: #dedede;
-    font-size: 0.8rem;
-}
-/*swiper.js*/
-.swiper-container {
-    width: 100%;
-    height: 25rem;
-    margin: 0 auto;
-    list-style: none;
-    padding: 0;
-}
-.swiper-container-vertical > .swiper-wrapper {
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -webkit-flex-direction: column;
-    -ms-flex-direction: column;
-    flex-direction: column;
-}
-.swiper-wrapper {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    display: -webkit-box;
-    display: -webkit-flex;
-    display: -ms-flexbox;
-    display: flex;
-    transition-property: transform;
-    box-sizing: content-box;
-}
-.swiper-container-multirow > .swiper-wrapper {
-    -webkit-flex-wrap: wrap;
-    -ms-flex-wrap: wrap;
-    flex-wrap: wrap;
-}
-.swiper-container-free-mode > .swiper-wrapper {
-    transition-timing-function: ease-out;
-    margin: 0 auto;
-}
-.swiper-slide .container {z-index: 1;}
-.swiper-slide:after {
-    content: '';
-    position: absolute;
-    left: 0; top: 0;
-    width: 100%; height: 100%;
-    background: #000;
-    opacity: .3;
-    z-index: -1;
-}
-/* Auto Height */
-.swiper-container-autoheight,
-.swiper-container-autoheight .swiper-slide {
-    height: auto;
-}
-.swiper-container-autoheight .swiper-wrapper {
-    -webkit-box-align: start;
-    -webkit-align-items: flex-start;
-    -ms-flex-align: start;
-    align-items: flex-start;
-    transition-property: height, -webkit-transform;
-    transition-property: transform, height;
-}
-.swiper-button-prev,
-.swiper-container-rtl .swiper-button-next {
-    color: white;
-    left: 10px;
-    right: auto;
-}
-.swiper-button-next,
-.swiper-container-rtl .swiper-button-prev {
-    color: white;
-    right: 10px;
-    left: auto;
-}
-.swiper-button-lock {
-    display: none;
-}
-.swiper-pagination {
-    position: absolute;
-    text-align: center;
-}
-/* Common Styles */
-.swiper-pagination-fraction,
-.swiper-pagination-custom,
-.swiper-container-horizontal > .swiper-pagination-bullets {
-    bottom: 10px;
-    left: 0;
-    width: 100%;
-}
- /*Bullets */
-.swiper-pagination-bullet {
-    width: 10px;
-    height: 10px;
-    display: inline-block;
-    border-radius: 100%;
-    background: #ccc;
-    opacity: 0.2;
-}
-.swiper-pagination-bullet-active {
-    opacity: 1;
-    background: #fff;
-}
-/*swiper img*/
-.img-1{
-    max-width: 100%;
-    max-height: 100%;
-    height: auto !important;
-    padding-left: 40px;
-    margin: 25px;
-}
-.promotion-1{
-    padding-top: 7rem;
-    margin-left: 1rem;
-}
-.img-2{
-    max-width: 100%;
-    max-height: 100%;
-    height: auto !important;
-    padding-left: 40px;
-    margin: 25px;
-}
-.promotion-2{
-    padding-top: 7rem;
-    margin-left: 1rem;
-}
-.img-3{
-    max-width: 100%;
-    max-height: 100%;
-    height: auto !important;
-    padding-left: 40px;
-    margin: 25px;
-}
-.promotion-3{
-    padding-top: 7rem;
-    margin-left: 1rem;
-}
