--- app/templates/layout/footer.html
@@ -1,9 +1,0 @@
-<div class="container my-10">
-    <nav class="navbar fixed-bottom navbar-light bg-light py-1">
-        <div id="footer">
-            <div class="container text-black">
-              <p class="footer-block">Sticky Footer</p>
-            </div>
-        </div>
-    </nav>
-</div>
