--- app/templates/layout/header.html
@@ -1,76 +1,0 @@
-<link rel="stylesheet" type="text/css" href="/static/css/header.css">
-<div class="container-fluid outer-1">
-    <div class="container-fluid d-grid align-items-center" style="grid-template-columns: 0.5fr 2fr; padding-top: 10px;">
-        <div class="dropdown">
-            <a class="logo" href="/main">Algo Camping</a>
-        </div>
-
-        <div class="d-flex align-items-center">
-            <div class="container">
-                <div class="row">
-                    <form class="align-items-center">
-                        <span>
-                            <button class="btn btn-secondary dropdown-toggle" type="button" id="default-menu" data-bs-toggle="dropdown" aria-expanded="false">
-                             전체
-                            </button>
-                            <ul class="dropdown-menu" id="area-menu">
-                                <li><a class="dropdown-item" href="#">서울</a></li>
-                                <li><a class="dropdown-item" href="#">경기</a></li>
-                                <li><a class="dropdown-item" href="#">인천</a></li>
-                                <li><a class="dropdown-item" href="#">강원</a></li>
-                                <li><a class="dropdown-item" href="#">충남</a></li>
-                                <li><a class="dropdown-item" href="#">대전</a></li>
-                                <li><a class="dropdown-item" href="#">충북</a></li>
-                                <li><a class="dropdown-item" href="#">부산</a></li>
-                                <li><a class="dropdown-item" href="#">울산</a></li>
-                                <li><a class="dropdown-item" href="#">대구</a></li>
-                                <li><a class="dropdown-item" href="#">경북</a></li>
-                                <li><a class="dropdown-item" href="#">경남</a></li>
-                                <li><a class="dropdown-item" href="#">전남</a></li>
-                                <li><a class="dropdown-item" href="#">광주</a></li>
-                                <li><a class="dropdown-item" href="#">전북</a></li>
-                                <li><a class="dropdown-item" href="#">제주</a></li>
-                            </ul>
-                        </span>
-                        <input class="col-12" type="search" data-role="tagsinput" value="" style="display: none">
-                        <button class="col-0 search-btn" type="button" style="">
-                            <i class="material-icons search" width="32" height="32">search</i>
-                        </button>
-                    </form>
-                </div>
-            </div>
-
-            <div class="flex-shrink-0 dropdown user-dropdown">
-                {% if param %}
-                    <a href="#" class="d-block link-dark text-decoration-none" id="dropdownUser2" data-bs-toggle="dropdown" aria-expanded="false">
-                        <i class="material-icons user" width="32" height="32" style="color: #1f284a;">account_circle</i>
-                        <p class="mx-2 py-1 username">{{ param.name }}님 안녕하세요</p>
-                    </a>
-                    <ul class="dropdown-menu text-small shadow" id="logout-menu">
-                        <li><a class="dropdown-item" id="logout-btn" href="#">로그아웃</a></li>
-                    </ul>
-                {% else %}
-                    <a href="#" class="d-block link-dark text-decoration-none" id="dropdownUser2" data-bs-toggle="dropdown" aria-expanded="false">
-                        <i class="material-icons user" width="32" height="32" style="color: #1f284a;">account_circle</i>
-                        <p class="mx-2 py-1 username">로그인</p>
-                    </a>
-                    <ul class="dropdown-menu text-small shadow" id="login-menu">
-                        <li><a class="dropdown-item" id="login-btn" href="/signin">로그인</a></li>
-                    </ul>
-                {% endif %}
-            </div>
-        </div>
-    </div>
-    <div class="row">
-        <span class="text-secondary text-center info-text">태그 혹은 캠핑장 명을 입력하세요. ex) 친절한, 알고캠핑장</span>
-    </div>
-</div>
-<script src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/typeahead.js/0.11.1/typeahead.bundle.min.js"></script>
-<script src="/static/js/tagsinput.js"></script>
-<script src="/static/css/header.css"></script>
-<script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
-<script>
-    Kakao.init('30d72d58b0b7ddc0a4ebd7fba31f8052');
-</script>
