--- app/templates/signup.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Algo Camping</title>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.0.1/css/bootstrap.min.css">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" type="text/css" href="/static/css/signup.css">
-    <link rel="shorcut icon" href="../static/imgs/favicon.ico" type="image/x-icon" sizes="16x16">
-</head>
-<body>
-    <div class="container">
-        <nav class="navbar fixed-top navbar-light bg-light p-4">
-            <div class="container-fluid">
-                <a class="navbar-brand col-md-12" href="/">
-                    <span class="navbar-text" id="logo">
-                        Algo Camping
-                    </span>
-                </a>
-            </div>
-        </nav>
-    </div>
-    <div class="container col-md-5 login">
-        <div class="row justify-content-md-center"></div>
-            <div class="p-5 bg-light col-md-auto rounded-3 outer-2">
-                <div class="container-fluid">
-                    <h2 class="border-bottom display-12 fw-bold py-2">Join</h2>
-                    <br>
-                    <div class="row display-12">
-                        <div class="mb-2">
-                            <label for="exampleDropdownFormEmail2" class="form-label">이메일</label>
-                            <div class="input-group mb-3">
-                                <input type="text" class="form-control" id="email-form" placeholder="email@example.com" arial-label="" aria-describedby="button-addon2">
-                                <button class="btn col-md-2" type="button" id="check-btn">확인</button>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="mb-3">
-                        <label for="exampleDropdownFormPassword2" class="form-label">패스워드</label>
-                        <input type="password" class="form-control" id="password-form" placeholder="Password">
-                    </div>
-                    <div class="mb-3">
-                        <label for="exampleDropdownFormPassword2" class="form-label">이름</label>
-                        <input type="text" class="form-control" id="name-form" placeholder="Name">
-                    </div>
-                    <button type="button" class="btn col-md-2 btn-signin py-2" id="signin-btn">OK</button><br>
-                </div><br>
-            </div><br>
-        </div>
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-<script src="/static/js/signup.js"></script>
-</body>
-</html>
