+++ camping_server1/app/main/__init__.py
@@ -2,7 +2,7 @@ from flask import Flask
 from ..view import routes
 from app.main.controller import user
 import warnings
-from ..config import DBConfig
+from ..config import DBConfig, Config
 from .model import db, migrate
 from flask_jwt_extended import *
 
@@ -18,6 +18,11 @@ def create_app():
     app.config.from_object(DBConfig)
     db.init_app(app)
     migrate.init_app(app, db)
+
+    app.config.update(
+        DEBUG=True,
+        JWT_SECRET_KEY=Config.JWT_SECRET_KEY
+    )
 
     jwt = JWTManager(app)
