+++ camping_server1/app/main/controller/user.py
@@ -1,4 +1,5 @@
 from app.main.service import user as user_service
+from app.main.util.user_dto import UserDTO as user_dto
 from flask_restx import Resource, Namespace
 from flask import jsonify, request, session, redirect
 from flask import Blueprint
@@ -6,6 +7,31 @@ from flask import Blueprint
 user = Namespace('user', description='relating to user')
 auth = Blueprint('auth', __name__, url_prefix='/auth')
 
+@user.route('/validation', methods=['POST'])
+class UserValidation(Resource):
+    def post(self):
+        """토큰 유효성 체크"""
+        param = {}
+
+        values = dict(request.values)
+        token = values['access_token']
+        try:
+            if token == session['access_token']:
+                param['code'] = 200
+            else:
+                # 비정상적인 접근
+                param['code'] = 403
+        except:
+            # 세션에 토큰이 만료되었거나 클라이언트에서 보낸 토큰 값이 잘못된 경우
+            try:
+                # getter
+                param = user_dto.user
+                user_service.delete_token(param['name'])
+            except:
+                pass
+            finally:
+                param['code'] = 403
+        return param
 
 @user.route('/check', methods=['POST'])
 class UserCheck(Resource):
@@ -14,14 +40,12 @@ class UserCheck(Resource):
         values = dict(request.values)
         return user_service.is_duplicate(values)
 
-
 @user.route('/signup', methods=['POST'])
 class UserSignup(Resource):
     def post(self):
         """회원가입"""
         values = dict(request.values)
         return user_service.signup(values)
-
 
 @user.route('/signin', methods=['POST'])
 class UserSignin(Resource):
@@ -30,14 +54,11 @@ class UserSignin(Resource):
         values = dict(request.values)
         return user_service.signin(values)
 
-
 @user.route('/detail', methods=['POST'])
-# @jwt_required()
 class UserDetail(Resource):
     def post(self):
         """사용자 정보 확인"""
         pass
-
 
 @user.route('/sns/signin', methods=['POST'])
 class UserSNSSignin(Resource):
@@ -57,16 +78,18 @@ class UserSNSSignin(Resource):
         session['id'] = id
         return jsonify({'code': 200})
 
-
 @auth.route('/signout')
 def signout():
     """사용자 로그아웃"""
-    user_service.delete_token(session['name'])
+    # getter
+    param = user_dto.user
+    user_service.delete_token(param['name'])
     return redirect(request.host_url, code=302)
-
 
 @auth.route('/sns/signout')
 def sns_signout():
     """플랫폼 로그아웃"""
-    user_service.delete_token(session['name'])
+    # getter
+    param = user_dto.user
+    user_service.delete_token(param['name'])
     return redirect(request.host_url, code=302)
