--- run.sh
@@ -1,3 +1,0 @@
-#export FLASK_APP=app.py
-#export FLASK_DEBUG=1
-#flask run
